Add functionality to save score to firebase
